pyproject.toml
[tool.poetry]
name = "table-gen"
version = "0.1.0"
description = ""
authors = ["eddie <eddie.zhang@plato.systems>"]

[tool.poetry.dependencies]
python = ">=3.6.2,<4.0"
Pillow = "^8.3.1"
google-api-python-client = "^2.15.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.5"

[tool.poetry.dev-dependencies]
black = {version = "^21.7b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
